working on some 64bit compatibility issues

+++ Source/DREAM3DLib/ReconstructionFilters/SegmentGrains.h
@@ -88,8 +88,8 @@ class DREAM3DLib_EXPORT SegmentGrains : public AbstractFilter
     virtual void execute();
     virtual void preflight();
 
-    virtual int getSeed(size_t gnum);
-    virtual bool determineGrouping(int referencepoint, int neighborpoint, size_t gnum);
+    virtual int64_t getSeed(size_t gnum);
+    virtual bool determineGrouping(int64_t referencepoint, int64_t neighborpoint, size_t gnum);
 
   protected:
     SegmentGrains();
